+++ source/module_elecstate/test/elecstate_base_test.cpp
@@ -98,8 +98,6 @@ void Charge::check_rho()
  *      - determine the number of electrons for spin up and down
  *   - Constructor: elecstate::ElecState(charge, rhopw, bigpw)
  *      - constructor ElecState using existing charge, rhopw, bigpw
- *   - CalNbands: elecstate::ElecState::cal_nbands()
- *      - calculate the number of bands
  *   - InitKS: elecstate::ElecState::init_ks()
  *      - initialize the elecstate for KS-DFT
  *   - GetRho: elecstate::ElecState::getRho()
@@ -181,145 +179,6 @@ TEST_F(ElecStateTest, Constructor)
     delete bigpw;
     delete rhopw;
     delete charge;
-}
-
-TEST_F(ElecStateTest, CalNbands)
-{
-    cal_nbands(GlobalV::nelec, GlobalV::NLOCAL, elecstate->nelec_spin, GlobalV::NBANDS);
-    EXPECT_EQ(GlobalV::NBANDS, 6);
-}
-
-TEST_F(ElecStateTest, CalNbandsFractionElec)
-{
-    GlobalV::nelec = 9.5;
-    cal_nbands(GlobalV::nelec, GlobalV::NLOCAL, elecstate->nelec_spin, GlobalV::NBANDS);
-    EXPECT_EQ(GlobalV::NBANDS, 6);
-}
-
-TEST_F(ElecStateTest, CalNbandsSOC)
-{
-    PARAM.input.lspinorb = true;
-    GlobalV::NBANDS = 0;
-    cal_nbands(GlobalV::nelec, GlobalV::NLOCAL, elecstate->nelec_spin, GlobalV::NBANDS);
-    EXPECT_EQ(GlobalV::NBANDS, 20);
-}
-
-TEST_F(ElecStateTest, CalNbandsSDFT)
-{
-    PARAM.input.esolver_type = "sdft";
-    EXPECT_NO_THROW(cal_nbands(GlobalV::nelec, GlobalV::NLOCAL, elecstate->nelec_spin, GlobalV::NBANDS));
-}
-
-TEST_F(ElecStateTest, CalNbandsLCAO)
-{
-    PARAM.input.basis_type = "lcao";
-    EXPECT_NO_THROW(cal_nbands(GlobalV::nelec, GlobalV::NLOCAL, elecstate->nelec_spin, GlobalV::NBANDS));
-}
-
-TEST_F(ElecStateDeathTest, CalNbandsLCAOINPW)
-{
-    PARAM.input.basis_type = "lcao_in_pw";
-    GlobalV::NLOCAL = GlobalV::NBANDS - 1;
-    testing::internal::CaptureStdout();
-    EXPECT_EXIT(cal_nbands(GlobalV::nelec, GlobalV::NLOCAL, elecstate->nelec_spin, GlobalV::NBANDS), ::testing::ExitedWithCode(0), "");
-    output = testing::internal::GetCapturedStdout();
-    EXPECT_THAT(output, testing::HasSubstr("NLOCAL < NBANDS"));
-}
-
-TEST_F(ElecStateDeathTest, CalNbandsWarning1)
-{
-    GlobalV::NBANDS = GlobalV::nelec / 2 - 1;
-    testing::internal::CaptureStdout();
-    EXPECT_EXIT(cal_nbands(GlobalV::nelec, GlobalV::NLOCAL, elecstate->nelec_spin, GlobalV::NBANDS), ::testing::ExitedWithCode(0), "");
-    output = testing::internal::GetCapturedStdout();
-    EXPECT_THAT(output, testing::HasSubstr("Too few bands!"));
-}
-
-TEST_F(ElecStateDeathTest, CalNbandsWarning2)
-{
-    GlobalV::NSPIN = 2;
-    GlobalV::nupdown = 4.0;
-    elecstate->init_nelec_spin();
-    testing::internal::CaptureStdout();
-    EXPECT_EXIT(cal_nbands(GlobalV::nelec, GlobalV::NLOCAL, elecstate->nelec_spin, GlobalV::NBANDS), ::testing::ExitedWithCode(0), "");
-    output = testing::internal::GetCapturedStdout();
-    EXPECT_THAT(output, testing::HasSubstr("Too few spin up bands!"));
-}
-
-TEST_F(ElecStateDeathTest, CalNbandsWarning3)
-{
-    GlobalV::NSPIN = 2;
-    GlobalV::nupdown = -4.0;
-    elecstate->init_nelec_spin();
-    testing::internal::CaptureStdout();
-    EXPECT_EXIT(cal_nbands(GlobalV::nelec, GlobalV::NLOCAL, elecstate->nelec_spin, GlobalV::NBANDS), ::testing::ExitedWithCode(0), "");
-    output = testing::internal::GetCapturedStdout();
-    EXPECT_THAT(output, testing::HasSubstr("Too few spin down bands!"));
-}
-
-TEST_F(ElecStateTest, CalNbandsSpin1)
-{
-    GlobalV::NSPIN = 1;
-    GlobalV::NBANDS = 0;
-    cal_nbands(GlobalV::nelec, GlobalV::NLOCAL, elecstate->nelec_spin, GlobalV::NBANDS);
-    EXPECT_EQ(GlobalV::NBANDS, 15);
-}
-
-TEST_F(ElecStateTest, CalNbandsSpin1LCAO)
-{
-    GlobalV::NSPIN = 1;
-    GlobalV::NBANDS = 0;
-    PARAM.input.basis_type = "lcao";
-    cal_nbands(GlobalV::nelec, GlobalV::NLOCAL, elecstate->nelec_spin, GlobalV::NBANDS);
-    EXPECT_EQ(GlobalV::NBANDS, 6);
-}
-
-TEST_F(ElecStateTest, CalNbandsSpin4)
-{
-    GlobalV::NSPIN = 4;
-    GlobalV::NBANDS = 0;
-    cal_nbands(GlobalV::nelec, GlobalV::NLOCAL, elecstate->nelec_spin, GlobalV::NBANDS);
-    EXPECT_EQ(GlobalV::NBANDS, 30);
-}
-
-TEST_F(ElecStateTest, CalNbandsSpin4LCAO)
-{
-    GlobalV::NSPIN = 4;
-    GlobalV::NBANDS = 0;
-    PARAM.input.basis_type = "lcao";
-    cal_nbands(GlobalV::nelec, GlobalV::NLOCAL, elecstate->nelec_spin, GlobalV::NBANDS);
-    EXPECT_EQ(GlobalV::NBANDS, 6);
-}
-
-TEST_F(ElecStateTest, CalNbandsSpin2)
-{
-    GlobalV::NSPIN = 2;
-    GlobalV::NBANDS = 0;
-    elecstate->init_nelec_spin();
-    cal_nbands(GlobalV::nelec, GlobalV::NLOCAL, elecstate->nelec_spin, GlobalV::NBANDS);
-    EXPECT_EQ(GlobalV::NBANDS, 16);
-}
-
-TEST_F(ElecStateTest, CalNbandsSpin2LCAO)
-{
-    GlobalV::NSPIN = 2;
-    GlobalV::NBANDS = 0;
-    PARAM.input.basis_type = "lcao";
-    elecstate->init_nelec_spin();
-    cal_nbands(GlobalV::nelec, GlobalV::NLOCAL, elecstate->nelec_spin, GlobalV::NBANDS);
-    EXPECT_EQ(GlobalV::NBANDS, 6);
-}
-
-TEST_F(ElecStateDeathTest, CalNbandsGaussWarning)
-{
-    Occupy::use_gaussian_broadening = true;
-    EXPECT_TRUE(Occupy::gauss());
-    GlobalV::NBANDS = 5;
-    testing::internal::CaptureStdout();
-    EXPECT_EXIT(cal_nbands(GlobalV::nelec, GlobalV::NLOCAL, elecstate->nelec_spin, GlobalV::NBANDS), ::testing::ExitedWithCode(0), "");
-    output = testing::internal::GetCapturedStdout();
-    EXPECT_THAT(output, testing::HasSubstr("for smearing, num. of bands > num. of occupied bands"));
-    Occupy::use_gaussian_broadening = false;
 }
 
 TEST_F(ElecStateTest, InitKS)
